+++ Changelog.md
@@ -11,3 +11,4 @@
 - Protected members used outside of inheritors are converted as public
 - Support conversion for annotation constructor calls
 - Place comments from the middle of the call to the end
+- Drop line breaks between operator arguments (except '+', "-", "&&" and "||")